Rename file to interfaces to match with #7998
