+++ sdk/core/core-tracing/src/interfaces.ts
